# docs/api/control-plane-openapi.yaml
openapi: 3.1.0
info:
  title: GuardMCP Control Plane Public API
  version: 1.0.0
  description: >-
    Public API v1 served by services/control-plane for the GuardMCP console.
    Most identifiers and demo data mirror the deterministic seeds in
    infra/postgres/001-demo-seed.sql and infra/redis/001-demo-seed.sh; the
    Replay-only fixtures used to exercise hash-chain and pagination behavior
    (see DemoSeed.SESSION_BROKEN_CHAIN_ID / SESSION_LARGE_ID) are in-memory
    only and intentionally absent from those seed files.
servers:
  - url: http://localhost:8080
    description: Local development (docker compose exposes the same port)
...
paths:
...
  /api/v1/servers/{id}:
    get:
      tags: [servers]
      summary: Server detail
      operationId: getServer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Full server entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpServer"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/servers/{id}/trust:
    put:
      tags: [servers]
      summary: Change a server's trust grade
      description: >-
        Grade order is untrusted < limited < trusted. A downgrade applies
        immediately regardless of `confirmed`. An upgrade without
        `confirmed: true` is rejected with 409 and an impact summary instead
        of applying; the same request with `confirmed: true` then applies it.
        A request for the server's current grade is a no-op. Every actual
        change is recorded as a TrustLevelChangeEvent in the audit hash chain
        and pushed to /api/v1/servers/stream subscribers immediately. This is
        the endpoint GMCP-80 §3.8.1's `PUT /servers/{id}` describes — the
        console (lib/api/client.ts) deliberately calls this trust
        sub-resource rather than a bare PUT /servers/{id}, so no separate
        endpoint was added.
      operationId: changeServerTrust
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrustChangeRequest"
      responses:
        "200":
          description: Updated (or unchanged, for a no-op) server summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerSummary"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: >-
            The requested change is an upgrade and confirmed was not true.
            details carries fromTrust, toTrust, and affectedPolicyCount.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  responses:
    BadRequest:
      description: The request is malformed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: The referenced resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable reason code
        message:
          type: string
        details:
          type: object
          additionalProperties:
            type: string
...
    TrustLevel:
      type: string
      enum: [trusted, limited, untrusted]
    ConnectionStatus:
      type: string
      enum: [connected, disconnected]
    ToolRisk:
      type: string
      enum: [high, medium, low]
    ToolSnapshotStatusView:
      type: object
      required: [state, pendingDiffCount]
      description: >-
        FR-GW-03 §6.1. Postgres-backed drift state, unlike ToolSummary's risk/policies which
        are still a static demo seed (see ToolSummary's own description).
      properties:
        state:
          type: string
          enum: [unapproved, in_sync, drift_detected, drift_acknowledged]
        snapshotCapturedAt:
          type: [string, "null"]
          format: date-time
        lastCheckedAt:
          type: [string, "null"]
          format: date-time
        pendingDiffCount:
          type: integer
        latestDiffId:
          type: [string, "null"]
          format: uuid
    ToolSummary:
      type: object
      required: [name, risk, policies, snapshotStatus]
      description: >-
        GMCP-80 §3.1. Field names match apps/console/lib/api/types.ts's ToolEntry
        (risk, policies) rather than the spec doc's illustrative riskLevel/policyIds.
        risk/policies are a fixed demo seed (no real per-tool risk scoring or
        policy binding exists yet); snapshotStatus is FR-GW-03's real, Postgres-backed
        drift state.
      properties:
        name:
          type: string
        risk:
          $ref: "#/components/schemas/ToolRisk"
        policies:
          type: array
          items:
            type: string
        snapshotStatus:
          $ref: "#/components/schemas/ToolSnapshotStatusView"
    ServerSummary:
      type: object
      required: [id, name, connected, trust, tools, endpoint]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        connected:
          type: boolean
        trust:
          $ref: "#/components/schemas/TrustLevel"
        tools:
          type: array
          items:
            $ref: "#/components/schemas/ToolSummary"
        endpoint:
          type: string
...
    McpServer:
      type: object
      required:
        - id
        - name
        - endpoint
        - trustLevel
        - connectionStatus
        - toolSnapshotHash
        - createdAt
        - updatedAt
        - trustLevelUpdatedAt
        - trustLevelUpdatedBy
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        endpoint:
          type: string
        trustLevel:
          $ref: "#/components/schemas/TrustLevel"
        connectionStatus:
          $ref: "#/components/schemas/ConnectionStatus"
        toolSnapshotHash:
          description: FR-GW-03 tool-snapshot hash; null until a first-approval snapshot exists.
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        trustLevelUpdatedAt:
          type: string
          format: date-time
        trustLevelUpdatedBy:
          description: 'null until a console identity system exists; "console" for a confirmed upgrade today.'
          type: [string, "null"]
    TrustChangeRequest:
      type: object
      required: [trustLevel]
      properties:
        trustLevel:
          $ref: "#/components/schemas/TrustLevel"
        confirmed:
          type: boolean
          default: false
...